+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "udata-front",
-  "version": "3.5.1.dev",
+  "version": "4.0.4",
   "description": "Open Data Portal - French customization",
   "author": {
     "name": "Etalab"
@@ -22,6 +22,7 @@
     "local-testing": "start-server-and-test prepare-and-serve-e2e http://localhost:7000 cypress",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
+    "test-storybook": "test-storybook --no-cache",
     "chromatic": "npx chromatic"
   },
   "overrides": {
@@ -54,6 +55,7 @@
     "@storybook/addon-styling": "^1.3.7",
     "@storybook/blocks": "^7.5.2",
     "@storybook/test": "^8.0.6",
+    "@storybook/test-runner": "^0.18.2",
     "@storybook/testing-library": "^0.2.2",
     "@storybook/vue3": "^7.5.2",
     "@storybook/vue3-vite": "^7.5.2",
@@ -66,6 +68,7 @@
     "axios": "^1.6.1",
     "browserslist": "^4.21.5",
     "chart.js": "^4.2.1",
+    "chromatic": "^11.5.1",
     "clipboard": "^2.0.11",
     "cypress": "^13.1.0",
     "cypress-axe": "^1.5.0",
