+++ udata_front/theme/gouvfr/assets/js/pages/OrganizationPublishingForm/Step2DescribeOrganization.stories.ts
@@ -4,7 +4,7 @@ import { within, waitFor, userEvent } from '@storybook/testing-library';
 import { expect } from '@storybook/test';
 import type { NewOrganization } from "@etalab/data.gouv.fr-components";
 import Step2DescribeOrganization from './Step2DescribeOrganization.vue';
-import * as Stepper from '../../components/Form/Stepper/Stepper.stories';
+import * as Step1 from './Step1CreateOrJoinOrganization.stories';
 
 
 const meta = {
@@ -31,7 +31,7 @@ const organization: NewOrganization = {
 
 const args = {
   organization,
-  steps: Stepper.StepperOrgaOnSecondStep.args.steps,
+  steps: Step1.Step1.args.steps,
   errors: []
 };
