+++ udata_front/theme/gouvfr/assets/js/pages/OrganizationPublishingForm/Step3CompleteTheOrganization.stories.ts
@@ -1,6 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/vue3';
 import Step3CompleteTheOrganization from './Step3CompleteTheOrganization.vue';
-import * as Stepper from '../../components/Form/Stepper/Stepper.stories';
 import * as Step2 from './Step2DescribeOrganization.stories';
 
 
@@ -18,7 +17,9 @@ organization.description = 'This is the provided description'
 
 const args = {
   organization,
-  steps: Stepper.StepperOrgaOnSecondStep.args.steps,
+  steps: Step2.Step2.args.steps,
+  datasetLink: "",
+  reuseLink: ""
 };
 
 export const Step3: StoryObj<typeof meta> = {
